Refactor: DTO refactoring
